requirements.txt
cairosvg==2.8.2
clip==1.0
diffusers==0.33.1
kagglehub==0.3.12
kmeans_gpu==0.0.5
lpips==0.1.4
matplotlib==3.10.3
more_itertools==8.10.0
numpy==2.2.6
opencv_python==4.11.0.86
pandas==2.2.3
peft==0.15.2
Pillow==11.2.1
torch==2.6.0+cu124
torchvision==0.21.0+cu124
transformers==4.52.2